feat(backend): - :zap: add middleware admin - :zap: add PriceController index and update - :zap: add FieldController index - :zap: add PriceFactory seeder - :zap: change price_pricing typedata to double - :zap: change total_payment typedata to double

+++ database/migrations/2022_05_11_125147_create_bookings_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->dateTime('booking_date');
             $table->timestamp('booking_hour');
             $table->string('payment_type');
-            $table->integer('total_payment');
+            $table->double('total_payment');
             $table->timestamps();
         });
     }
